installer/requirements.txt
click>=7.0
PyYAML
requests
psutil
six
virtualenv-api>=2.1.18
